+++ crates/icu/src/macros/start.rs
@@ -5,7 +5,10 @@ macro_rules! __icu_load_config {
         #[cfg(icu_config)]
         {
             let config_str = include_str!(env!("ICU_CONFIG_PATH"));
-            $crate::config::Config::init_from_toml(config_str).unwrap()
+            $crate::expect_or_trap(
+                $crate::config::Config::init_from_toml(config_str),
+                "init config",
+            )
         }
     };
 }
@@ -24,7 +27,10 @@ macro_rules! icu_start {
             // setup
             ::icu::ops::state::save_state(&bundle);
             ::icu::memory::CanisterState::set_parents(parents);
-            ::icu::memory::CanisterState::set_type(&$canister_type).unwrap();
+            $crate::expect_or_trap(
+                ::icu::memory::CanisterState::set_type(&$canister_type),
+                "set canister type",
+            );
             __icu_shared_setup();
 
             let _ = ::icu::cdk::timers::set_timer(::std::time::Duration::from_secs(0), move || {
@@ -65,7 +71,10 @@ macro_rules! icu_start_root {
             ::icu::log!(::icu::Log::Info, "🏁 init: root");
 
             // setup
-            ::icu::memory::CanisterState::set_type(&::icu::types::CanisterType::ROOT).unwrap();
+            $crate::expect_or_trap(
+                ::icu::memory::CanisterState::set_type(&::icu::types::CanisterType::ROOT),
+                "set canister type",
+            );
             __icu_shared_setup();
 
             // register in CanisterRegistry